Exercise "Starting in the top left corner of a 20×20 grid, and only b…

+++ Services/Interfaces/INumbersService.cs
@@ -8,7 +8,7 @@ public interface INumbersService
     public int FindSumOfDigitsOfAPositiveNumber(int number);
     public int FindSumOfDigitsOfAPositiveNumberParsingThroughEveryDigit(int number);
     public int FindSumOfDigitsOfAPositiveNumberUsingLINQ(int number);
-    public int FindFactorialOfAPositiveNumber(int number);
+    public double FindFactorialOfAPositiveNumber(int number);
     public int? FindFactorialOfAPositiveNumberUsingRecursion(int? number, int? factorialValue);
     public int FindFactorialOfAPositiveNumberUsingWhileLoop(int number);
     public int FibonacciSeriesCalculationAndDisplay(int numberOfElements);
